+++ src/main/java/com/unknown/bankapp/services/AtmBalanceService.java
@@ -3,9 +3,9 @@ package com.unknown.bankapp.services;
 import java.util.Currency;
 
 public interface AtmBalanceService {
-    Long showBalance();
+    Double showBalance();
 
     Currency loadCurrency();
 
-    void changeBalance(Long newBalance);
+    void changeBalance(Double newBalance);
 }
